+++ src/main/java/mod/gottsch/neo/gottschcore/block/HalfBlock.java
@@ -1,0 +1,38 @@
+
+package mod.gottsch.neo.gottschcore.block;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.Half;
+
+/**
+ * @author by Mark Gottschling on 5/7/2025
+ */
+public class HalfBlock extends Block implements IHalfBlock {
+
+    public HalfBlock(Properties properties) {
+        super(properties);
+        this.registerDefaultState(defaultBlockState()
+                .setValue(HALF, Half.BOTTOM));
+    }
+
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
+        super.createBlockStateDefinition(builder);
+        builder.add(HALF);
+    }
+
+    public BlockState getStateForPlacement(BlockPlaceContext context) {
+        Direction direction = context.getClickedFace();
+        BlockPos pos = context.getClickedPos();
+        return defaultBlockState().setValue(HALF, direction != Direction.DOWN && (direction == Direction.UP || !(context.getClickLocation().y - (double) pos.getY() > 0.5D)) ? Half.BOTTOM : Half.TOP);
+    }
+
+    @Override
+    public Half getHalf(BlockState state) {
+        return (Half) state.getValue(HALF);
+    }
+}