+++ CHANGELOG.md
@@ -1,9 +1,25 @@
-**# Changelog for GottschCore 1.20.1
+**# Changelog for GottschCore 1.21.1 for NeoForge
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [2.5.0] - 2025-06-15
+
+### Added
+
+- Interval implementation for bst package.
+
+## [2.4.0] - 2025-05-09
+
+### Added
+
+- IHalfBlock block interface
+- HalfBlock block
+- FacingHalfBlock
+- WaterloggedFacingHalfBlock block
+- ModUtil class - contains asLocation(), mcLocation() methods, spawnMob(), spawnAndAddMob() methods.
 
 ## [2.3.0] - 2024-10-27